# BITS/SE4ML/Group89_Assignment2/Source_bundle/evidence/architecture/openapi_schema.json
{
  "components": {
    "schemas": {
      "Body_analyze_complete_api_v1_analyze_complete_post": {
        "properties": {
          "image": {
            "format": "binary",
            "title": "Image",
            "type": "string"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "required": [
          "image"
        ],
        "title": "Body_analyze_complete_api_v1_analyze_complete_post",
        "type": "object"
      },
      "Body_analyze_image_api_v1_image_analyze_post": {
        "properties": {
          "image": {
            "format": "binary",
            "title": "Image",
            "type": "string"
          }
        },
        "required": [
          "image"
        ],
        "title": "Body_analyze_image_api_v1_image_analyze_post",
        "type": "object"
      },
      "Body_classify_image_api_v1_image_classify_post": {
        "properties": {
          "image": {
            "format": "binary",
            "title": "Image",
            "type": "string"
          }
        },
        "required": [
          "image"
        ],
        "title": "Body_classify_image_api_v1_image_classify_post",
        "type": "object"
      },
      "ClassificationResponse": {
        "additionalProperties": false,
        "description": "Complete multilabel classification response.",
        "properties": {
          "api_version": {
            "const": "v1",
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "crossed_finding_names": {
            "items": {
              "type": "string"
            },
            "title": "Crossed Finding Names",
            "type": "array"
          },
          "educational_use_only": {
            "const": true,
            "default": true,
            "title": "Educational Use Only",
            "type": "boolean"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingEvidence"
            },
            "maxItems": 14,
            "minItems": 14,
            "title": "Findings",
            "type": "array"
          },
          "image": {
            "$ref": "#/components/schemas/ImageMetadata"
          },
          "interpretation": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Interpretation",
            "type": "string"
          },
          "latency_ms": {
            "minimum": 0.0,
            "title": "Latency Ms",
            "type": "number"
          },
          "model_versions": {
            "$ref": "#/components/schemas/ModelVersions"
          },
          "no_target_finding": {
            "title": "No Target Finding",
            "type": "boolean"
          },
          "prediction_id": {
            "format": "uuid",
            "title": "Prediction Id",
            "type": "string"
          },
          "prompt_registry_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Registry Version"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "const": "success",
            "default": "success",
            "title": "Status",
            "type": "string"
          },
          "timestamp_utc": {
            "format": "date-time",
            "title": "Timestamp Utc",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "timestamp_utc",
          "model_versions",
          "latency_ms",
          "prediction_id",
          "image",
          "findings",
          "crossed_finding_names",
          "no_target_finding",
          "interpretation"
        ],
        "title": "ClassificationResponse",
        "type": "object"
      },
      "CompleteAnalysisResponse": {
        "additionalProperties": false,
        "description": "Complete image, visual, and guarded language workflow.",
        "properties": {
          "api_version": {
            "const": "v1",
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "crossed_finding_names": {
            "items": {
              "type": "string"
            },
            "title": "Crossed Finding Names",
            "type": "array"
          },
          "educational_use_only": {
            "const": true,
            "default": true,
            "title": "Educational Use Only",
            "type": "boolean"
          },
          "explainability": {
            "$ref": "#/components/schemas/ExplainabilityContract"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingEvidence"
            },
            "maxItems": 14,
            "minItems": 14,
            "title": "Findings",
            "type": "array"
          },
          "image": {
            "$ref": "#/components/schemas/ImageMetadata"
          },
          "interpretation": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Interpretation",
            "type": "string"
          },
          "language_outputs": {
            "items": {
              "$ref": "#/components/schemas/EmbeddedLanguageOutput"
            },
            "maxItems": 4,
            "minItems": 3,
            "title": "Language Outputs",
            "type": "array"
          },
          "latency_ms": {
            "minimum": 0.0,
            "title": "Latency Ms",
            "type": "number"
          },
          "model_versions": {
            "$ref": "#/components/schemas/ModelVersions"
          },
          "no_target_finding": {
            "title": "No Target Finding",
            "type": "boolean"
          },
          "prediction_id": {
            "format": "uuid",
            "title": "Prediction Id",
            "type": "string"
          },
          "prompt_registry_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Registry Version"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "const": "success",
            "default": "success",
            "title": "Status",
            "type": "string"
          },
          "timestamp_utc": {
            "format": "date-time",
            "title": "Timestamp Utc",
            "type": "string"
          },
          "visual_evidence": {
            "items": {
              "$ref": "#/components/schemas/GradCAMEvidence"
            },
            "title": "Visual Evidence",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "timestamp_utc",
          "model_versions",
          "latency_ms",
          "prediction_id",
          "image",
          "findings",
          "crossed_finding_names",
          "no_target_finding",
          "interpretation",
          "explainability",
          "visual_evidence",
          "language_outputs"
        ],
        "title": "CompleteAnalysisResponse",
        "type": "object"
      },
...
      "GroundedGenerationRequest": {
        "additionalProperties": false,
        "description": "Request a language task using an existing prediction context.",
        "properties": {
          "prediction_id": {
            "format": "uuid",
            "title": "Prediction Id",
            "type": "string"
          }
        },
        "required": [
          "prediction_id"
        ],
        "title": "GroundedGenerationRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImageAnalysisResponse": {
        "additionalProperties": false,
        "description": "Classification response extended with Grad-CAM evidence.",
        "properties": {
          "api_version": {
            "const": "v1",
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "crossed_finding_names": {
            "items": {
              "type": "string"
            },
            "title": "Crossed Finding Names",
            "type": "array"
          },
          "educational_use_only": {
            "const": true,
            "default": true,
            "title": "Educational Use Only",
            "type": "boolean"
          },
          "explainability": {
            "$ref": "#/components/schemas/ExplainabilityContract"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingEvidence"
            },
            "maxItems": 14,
            "minItems": 14,
            "title": "Findings",
            "type": "array"
          },
          "image": {
            "$ref": "#/components/schemas/ImageMetadata"
          },
          "interpretation": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Interpretation",
            "type": "string"
          },
          "latency_ms": {
            "minimum": 0.0,
            "title": "Latency Ms",
            "type": "number"
          },
          "model_versions": {
            "$ref": "#/components/schemas/ModelVersions"
          },
          "no_target_finding": {
            "title": "No Target Finding",
            "type": "boolean"
          },
          "prediction_id": {
            "format": "uuid",
            "title": "Prediction Id",
            "type": "string"
          },
          "prompt_registry_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Registry Version"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "const": "success",
            "default": "success",
            "title": "Status",
            "type": "string"
          },
          "timestamp_utc": {
            "format": "date-time",
            "title": "Timestamp Utc",
            "type": "string"
          },
          "visual_evidence": {
            "items": {
              "$ref": "#/components/schemas/GradCAMEvidence"
            },
            "title": "Visual Evidence",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "timestamp_utc",
          "model_versions",
          "latency_ms",
          "prediction_id",
          "image",
          "findings",
          "crossed_finding_names",
          "no_target_finding",
          "interpretation",
          "explainability",
          "visual_evidence"
        ],
        "title": "ImageAnalysisResponse",
        "type": "object"
      },
...
      "LanguageGenerationResponse": {
        "additionalProperties": false,
        "description": "Guarded output from one grounded language task.",
        "properties": {
          "api_version": {
            "const": "v1",
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "educational_use_only": {
            "const": true,
            "default": true,
            "title": "Educational Use Only",
            "type": "boolean"
          },
          "generated_tokens": {
            "minimum": 1.0,
            "title": "Generated Tokens",
            "type": "integer"
          },
          "generation_latency_ms": {
            "minimum": 0.0,
            "title": "Generation Latency Ms",
            "type": "number"
          },
          "grounded_finding_names": {
            "items": {
              "type": "string"
            },
            "title": "Grounded Finding Names",
            "type": "array"
          },
          "guardrail_action": {
            "enum": [
              "accepted_model_generation",
              "safe_template_fallback"
            ],
            "title": "Guardrail Action",
            "type": "string"
          },
          "latency_ms": {
            "minimum": 0.0,
            "title": "Latency Ms",
            "type": "number"
          },
          "model_versions": {
            "$ref": "#/components/schemas/ModelVersions"
          },
          "no_target_finding": {
            "title": "No Target Finding",
            "type": "boolean"
          },
          "output_text": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Output Text",
            "type": "string"
          },
          "prediction_id": {
            "format": "uuid",
            "title": "Prediction Id",
            "type": "string"
          },
          "prompt_registry_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Registry Version"
          },
          "question": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "const": "success",
            "default": "success",
            "title": "Status",
            "type": "string"
          },
          "task_type": {
            "enum": [
              "structured_report",
              "plain_language_explanation",
              "grounded_question_answering",
              "educational_follow_up"
            ],
            "title": "Task Type",
            "type": "string"
          },
          "timestamp_utc": {
            "format": "date-time",
            "title": "Timestamp Utc",
            "type": "string"
          },
          "trigger_reasons": {
            "items": {
              "enum": [
                "task_routing_issue",
                "section_order_issue",
                "missing_required_finding_issue",
                "unsupported_finding_issue",
                "numeric_grounding_issue",
                "safety_boundary_issue",
                "no_target_boundary_issue",
                "qa_refusal_issue",
                "gradcam_boundary_issue",
                "forbidden_claim_issue"
              ],
              "type": "string"
            },
            "title": "Trigger Reasons",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "timestamp_utc",
          "model_versions",
          "latency_ms",
          "prediction_id",
          "task_type",
          "grounded_finding_names",
          "no_target_finding",
          "output_text",
          "guardrail_action",
          "generated_tokens",
          "generation_latency_ms"
        ],
        "title": "LanguageGenerationResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Educational decision-support API for frozen ChestMNIST classification, visual evidence, and grounded language generation. This service is not a diagnostic system.",
    "title": "API-Driven Chest X-Ray Analysis and Explanation Assistant",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analyze-complete": {
      "post": {
        "description": "Run classification, Grad-CAM, and guarded language generation.",
        "operationId": "analyze_complete_api_v1_analyze_complete_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_complete_api_v1_analyze_complete_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run the complete chest X-ray analysis workflow",
        "tags": [
          "complete-analysis"
        ]
      }
    },
    "/api/v1/explanation/generate": {
      "post": {
        "description": "Explain stored model output using guarded simple language.",
        "operationId": "generate_explanation_api_v1_explanation_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroundedGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageGenerationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate a grounded plain-language explanation",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/follow-up/recommend": {
      "post": {
        "description": "Generate non-diagnostic educational follow-up guidance.",
        "operationId": "recommend_follow_up_api_v1_follow_up_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroundedGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageGenerationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate controlled educational follow-up",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/image/analyze": {
      "post": {
        "description": "Run classification and threshold-controlled Grad-CAM.",
        "operationId": "analyze_image_api_v1_image_analyze_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_image_api_v1_image_analyze_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Classify an image and generate visual evidence",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/image/classify": {
      "post": {
        "description": "Run frozen multilabel classification and store the prediction.",
        "operationId": "classify_image_api_v1_image_classify_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_classify_image_api_v1_image_classify_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Classify a chest X-ray image",
        "tags": [
          "analysis"
        ]
      }
    },
...
    "/api/v1/predictions/{prediction_id}": {
      "get": {
        "description": "Return one stored prediction and its attached outputs.",
        "operationId": "get_prediction_api_v1_predictions__prediction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "prediction_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Prediction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredPredictionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retrieve one stored prediction",
        "tags": [
          "complete-analysis"
        ]
      }
    },
    "/api/v1/question/answer": {
      "post": {
        "description": "Answer only from server-controlled stored grounding values.",
        "operationId": "answer_grounded_question_api_v1_question_answer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroundedQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageGenerationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer a question using stored model output",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/report/generate": {
      "post": {
        "description": "Generate a guarded structured report from a stored prediction.",
        "operationId": "generate_report_api_v1_report_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroundedGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageGenerationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate a grounded preliminary model report",
        "tags": [
          "analysis"
        ]
      }
    },
...
  }
}